parallel sent of logs to graylog
